requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
attrs==23.2.0
bitarray==2.9.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cytoolz==0.12.2
eth-account==0.10.0
eth-hash==0.6.0
eth-keyfile==0.7.0
eth-keys==0.5.0
eth-rlp==1.0.0
eth-typing==4.0.0
eth-utils==3.0.0
eth_abi==5.0.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.41
hexbytes==0.3.1
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
numpy==1.26.3
packaging==23.2
pandas==2.1.4
parsimonious==0.9.0
pillow==10.2.0
protobuf==4.25.2
pyarrow==14.0.2
pycryptodome==3.20.0
pydeck==0.8.1b0
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyunormalize==15.1.0
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rlp==4.0.0
rpds-py==0.17.1
six==1.16.0
smmap==5.0.1
streamlit==1.30.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
web3==6.14.0
websockets==12.0
yarl==1.9.4
zipp==3.17.0
